Created class for fetching data, refactoring
