+++ src/api/Controller.js
@@ -1,6 +1,6 @@
 import Client from './Client';
 
-import { getProjectedCoords } from './util/calc';
+import { getProjectedCoords } from './calc';
 import { mapOverlays as colors } from '../config/colors.json';
 import {
     mapOverlays as constants,
